Start of key sharing stuff.
Adds button, which creates a file and launches a "Send email" intent
which prompts the user to send an email with the created file attached.

To send the file, because it's in internal storage, had to create a ContentProvider
so the email app can access it.

Also adds an intent-filter to the manifest so when you select a *.phnky attachment,
our app is listed as one of the "Complete this action with" apps.  Tested for Gmail
app, not sure if it'll work elsewhere (e.g. Yahoo).