Adds persistant key/pass/lockid storage.
KeyPassFileOpener is called to open .phnky email attachments.  It uses
FileManager to add the info to our internal file structure.

Our internal file structure is just a directory ("lockfiles") under
our app's internal directory (Context.getFilesDir()).  Each lock is stored
in a new file, starting with 1.phnky and counting up.  The file format is:

32 byte key, 1 byte passcode length, passcode length, 1 byte lock ID, lock ID.

Lock info is now kept neatly in a class, KeyPassData.  Eventually, this will
also be able to generate new keys and passcodes for initializing a lock.

On creation (and resume, I think), MainActivity will call FileManager.readData,
which reads all the *.phnky files, creates a KeyPassData object for each,
and returns an ArrayList of the objects.  As of now, MainActivity just selects the
first KeyPassData object in the list as the "currentKpd".